+++ src/openpkflow/be/__init__.py
@@ -3,6 +3,13 @@
 from __future__ import annotations
 
 from openpkflow.be.methods import BETOSTResult, be_sample_size, be_tost, be_tost_power
+from openpkflow.be.replicate import (
+    ReplicateBEResult,
+    cv_to_s_within,
+    ema_scaled_limits,
+    replicate_be,
+    s_within_to_cv_pct,
+)
 from openpkflow.be.results import BEResult
 from openpkflow.be.study import BEStudy
 
@@ -10,7 +17,12 @@ __all__ = [
     "BEStudy",
     "BEResult",
     "BETOSTResult",
+    "ReplicateBEResult",
     "be_tost",
     "be_tost_power",
     "be_sample_size",
+    "cv_to_s_within",
+    "ema_scaled_limits",
+    "replicate_be",
+    "s_within_to_cv_pct",
 ]